Add frontend for expense
* Permit /css

* Update when current user is determined

* Add expenses management pages and navigation updates

* Expand expense edit form to handle all fields

* Add delete controls to expense edit view

* Add expense detail view and clickable rows

* Normalize expense amount precision on edit form

* Add login and logout functionality

* Add UI Expense creation

+++ src/main/java/arobu/glitterfinv2/model/dto/ExpenseEntryApiPostDTO.java
@@ -1,7 +1,10 @@
 package arobu.glitterfinv2.model.dto;
 
+
+import java.math.BigDecimal;
+
 public class ExpenseEntryApiPostDTO {
-    private Double amount;
+    private BigDecimal amount;
     private String timestamp;
     private String source;
     private String merchant;
@@ -13,7 +16,7 @@ public class ExpenseEntryApiPostDTO {
     private Boolean shared;
     private Boolean outlier;
 
-    public Double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
